eapl_codebase/eapl_ws/requirements.txt
django-basicauth==0.5.2
django-rq==2.3.2
